pyproject.toml
[project]
name = "books-catalogue-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.122.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4,<1.8.0",
    "bcrypt==3.2.2",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "rich>=14.2.0",
    "sqlmodel>=0.0.30",
    "typer>=0.21.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
